Fix version check for Elixir
Prior to this patch, we couldn't use this package because it was
crashing when using Elixir pre-releases (here, 1.18.0-dev):

    == Compilation error in file lib/x509/logger.ex ==
    ** (ArgumentError) errors were found at the given arguments:

      * 1st argument: not a textual representation of an integer

        :erlang.binary_to_integer("0-dev")
        (elixir 1.18.0-dev) lib/enum.ex:1703: Enum."-map/2-lists^map/1-1-"/2
        (elixir 1.18.0-dev) lib/enum.ex:1703: Enum."-map/2-lists^map/1-1-"/2
        lib/x509/logger.ex:7: (module)

This is now fixed. `X509.Util.app_version/1` is still used in a couple
other places but just for applications that ship with OTP and they
should not have this problem.