Fix signature algorithm for certificates signed by EdDSA keys
Previously, when supplying an EdDSA key to functions in `X509.Certificate` that create certificates, the resulting
certificate would contain AlgorithmIdentifiers for ECDSA signatures (eg. `ecdsa-with-sha256`), despite containing an
EdDSA signature.

Instead, we now check for Edwards curves being used in an ECPrivateKey in `X509.SignatureAlgorithm.new`, and return a
proper AlgorithmIdentifier for that curve as specified by RFC 8410.